requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dateutil>=2.8
